+++ Changelog.md
@@ -4,7 +4,7 @@
 
 -   Option to program and evaluate the nRF Cloud multi-service firmware on the
     nRF9161 DK and the nRF9151 DK.
--   `SIM Card` step for nRF9151 DK.
+-   `SIM Card` step for nRF9151 DK and nRF9161 DK.
 
 ## 0.2.6 - 2024-04-12